requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
pdfplumber==0.10.3
python-dotenv==1.0.1
openai==1.12.0
google-generativeai==0.3.2
fpdf==1.7.2
Werkzeug==3.0.1
PyMySQL==1.1.0  # For MySQL database on PythonAnywhere
Jinja2==3.1.3
requests==2.31.0
gunicorn==21.2.0
streamlit
beautifulsoup4
pandas
transformers
torch
sentence-transformers
scikit-learn
googlemaps
pdfkit
psycopg2-binary==2.9.9
PyPDF2==3.0.1
python-docx==1.0.1
email-validator==2.1.0.post1 